refactor(harness-config): isolate typed settings patches
