+++ src/loushang/harness/config/agent/manager.py
@@ -2,21 +2,29 @@ from __future__ import annotations
 
 from collections.abc import Callable, Iterable, Mapping, Sequence
 from copy import deepcopy
-from dataclasses import asdict, dataclass, is_dataclass, replace
+from dataclasses import dataclass, replace
 from pathlib import Path
 from typing import Any, Literal, cast
 
 from loushang.agent import ThinkingLevel
 from loushang.ai.model import ModelSelection
 from loushang.harness.config import (
-    ConfigFieldSpec,
     ConfigLayer,
     LayeredConfig,
-    SchemaConfigCodec,
     ScopedConfigRuntime,
     SettingsRuntime,
-    decode_dataclass_patch,
-    encode_dataclass_diff,
+)
+from loushang.harness.config.agent._settings_codec import (
+    CONTROL_CONFIG_CODEC,
+    control_config_to_patch,
+    decode_package_source,
+)
+from loushang.harness.config.agent._settings_patch import (
+    UNSET,
+    AgentSettingsUpdate,
+    Unset,
+    build_settings_patch,
+    prepare_override_patch,
 )
 from loushang.harness.config.agent.types import (
     BranchSummarySettings,
@@ -24,7 +32,6 @@ from loushang.harness.config.agent.types import (
     ControlConfig,
     DoubleEscapeAction,
     ExternalToolPolicy,
-    HeadlessApprovalMode,
     ImageSettings,
     KeybindingValue,
     MarkdownSettings,
@@ -33,10 +40,7 @@ from loushang.harness.config.agent.types import (
     QueueMode,
     RetrySettings,
     SandboxSettings,
-    StatusLineAutoValue,
     StatusLineControlSettings,
-    StatusLineSeparator,
-    StatusLineStyle,
     TerminalSettings,
     ThinkingBudgetMap,
     ToolSettings,
@@ -61,16 +65,6 @@ SettingsScope = Literal["session", "global", "project"]
 ThinkingBudgetKey = Literal["minimal", "low", "medium", "high"]
 
 
-class _Unset:
-    __slots__ = ()
-
-
-_UNSET = _Unset()
-_REMOVED_SETTING_MESSAGES = {
-    "transport": "transport setting has been removed; use provider/contrib-specific configuration instead",
-}
-
-
 @dataclass(frozen=True)
 class SettingsError:
     scope: SettingsScope
@@ -78,683 +72,9 @@ class SettingsError:
     error: Exception
 
 
-def _normalize_string_sequence(
-    value: Iterable[str], field_name: str
-) -> tuple[str, ...]:
-    if isinstance(value, str):
-        raise TypeError(f"{field_name} must be a sequence of strings, not a string")
-    normalized: list[str] = []
-    for item in value:
-        if not isinstance(item, str):
-            raise TypeError(f"{field_name} must only contain strings")
-        normalized.append(item)
-    return tuple(normalized)
-
-
-def _normalize_package_source_sequence(
-    value: object, field_name: str = "packages"
-) -> tuple[PackageSourceConfig, ...]:
-    if not isinstance(value, Sequence) or isinstance(value, str):
-        raise TypeError(f"{field_name} must be a sequence")
-    return tuple(_deserialize_package_source(item) for item in value)
-
-
-def _deserialize_package_source(value: object) -> PackageSourceConfig:
-    if isinstance(value, PackageSourceConfig):
-        return value
-    if isinstance(value, str):
-        return PackageSourceConfig(source=value)
-    if not isinstance(value, Mapping):
-        raise TypeError("package source entries must be strings or objects")
-    source = value.get("source")
-    if not isinstance(source, str) or not source:
-        raise TypeError("package source object must include a non-empty string source")
-    return PackageSourceConfig(
-        source=source,
-        extensions=_optional_string_tuple(value.get("extensions"), "extensions"),
-        skills=_optional_string_tuple(value.get("skills"), "skills"),
-        prompts=_optional_string_tuple(value.get("prompts"), "prompts"),
-        themes=_optional_string_tuple(value.get("themes"), "themes"),
-    )
-
-
-def _optional_string_tuple(value: object, field_name: str) -> tuple[str, ...] | None:
-    if value is None:
-        return None
-    if not isinstance(value, Sequence) or isinstance(value, str):
-        raise TypeError(f"{field_name} must be a sequence of strings")
-    return _normalize_string_sequence(value, field_name)
-
-
-def _serialize_package_source(source: PackageSourceConfig) -> str | dict[str, object]:
-    if not source.filtered:
-        return source.source
-    payload: dict[str, object] = {"source": source.source}
-    if source.extensions is not None:
-        payload["extensions"] = list(source.extensions)
-    if source.skills is not None:
-        payload["skills"] = list(source.skills)
-    if source.prompts is not None:
-        payload["prompts"] = list(source.prompts)
-    if source.themes is not None:
-        payload["themes"] = list(source.themes)
-    return payload
-
-
-def _serialize_model_selection(
-    selection: ModelSelection | None,
-) -> dict[str, str] | None:
-    if selection is None:
-        return None
-    payload = {"provider": selection.provider, "model_id": selection.model_id}
-    if selection.endpoint_id:
-        payload["endpoint_id"] = selection.endpoint_id
-    return payload
-
-
-def _deserialize_model_selection(value: object) -> ModelSelection | None:
-    if value is None:
-        return None
-    if not isinstance(value, Mapping):
-        raise TypeError("default_model must be a JSON object or null")
-    provider = value.get("provider")
-    model_id = value.get("model_id")
-    if not isinstance(provider, str) or not isinstance(model_id, str):
-        raise TypeError(
-            "default_model must include string provider and model_id values"
-        )
-    endpoint_id = value.get("endpoint_id") or value.get("endpointId")
-    return ModelSelection(
-        provider=provider,
-        model_id=model_id,
-        endpoint_id=endpoint_id if isinstance(endpoint_id, str) else None,
-    )
-
-
-def _deserialize_queue_mode(value: object, field_name: str) -> QueueMode:
-    if value not in {"all", "one-at-a-time"}:
-        raise ValueError(f"{field_name} must be 'all' or 'one-at-a-time'")
-    return cast(QueueMode, value)
-
-
-def _deserialize_double_escape_action(value: object) -> DoubleEscapeAction:
-    if value not in {"fork", "tree", "none"}:
-        raise ValueError("double_escape_action must be 'fork', 'tree', or 'none'")
-    return cast(DoubleEscapeAction, value)
-
-
-def _deserialize_tree_filter_mode(value: object) -> TreeFilterMode:
-    if value not in {"default", "no-tools", "user-only", "labeled-only", "all"}:
-        raise ValueError(
-            "tree_filter_mode must be 'default', 'no-tools', 'user-only', 'labeled-only', or 'all'"
-        )
-    return cast(TreeFilterMode, value)
-
-
-def _deserialize_external_tool_policy(value: object) -> ExternalToolPolicy:
-    if value not in {"never", "auto", "required"}:
-        raise ValueError("external_tool_policy must be 'never', 'auto', or 'required'")
-    return cast(ExternalToolPolicy, value)
-
-
-def _deserialize_headless_approval_mode(value: object) -> HeadlessApprovalMode | None:
-    if value is None:
-        return None
-    if value not in {"allow", "deny"}:
-        raise ValueError("approval_mode must be 'allow', 'deny', or null")
-    return cast(HeadlessApprovalMode, value)
-
-
-def _deserialize_permission_profile(value: object) -> PermissionProfileId:
-    if not isinstance(value, str):
-        raise TypeError("permissions.profile must be a string")
-    permission_profile(value)
-    return cast(PermissionProfileId, value)
-
-
-def _deserialize_statusline_auto_value(
-    value: object, field_name: str
-) -> StatusLineAutoValue:
-    if value not in {"auto", "true", "false"}:
-        raise ValueError(f"{field_name} must be 'auto', 'true', or 'false'")
-    return cast(StatusLineAutoValue, value)
-
-
-def _deserialize_statusline_separator(
-    value: object, field_name: str
-) -> StatusLineSeparator:
-    if value not in {"pipe", "dot"}:
-        raise ValueError(f"{field_name} must be 'pipe' or 'dot'")
-    return cast(StatusLineSeparator, value)
-
-
-def _deserialize_statusline_style(value: object, field_name: str) -> StatusLineStyle:
-    if value not in {"codex-like", "muted", "plain"}:
-        raise ValueError(f"{field_name} must be 'codex-like', 'muted', or 'plain'")
-    return cast(StatusLineStyle, value)
-
-
-def _optional_string(value: object, field_name: str) -> str | None:
-    if value is None:
-        return None
-    if not isinstance(value, str):
-        raise TypeError(f"{field_name} must be a string or null")
-    return value
-
-
-def _bool_value(value: object, field_name: str) -> bool:
-    if not isinstance(value, bool):
-        raise TypeError(f"{field_name} must be a boolean")
-    return value
-
-
-def _string_tuple_or_none(value: object, field_name: str) -> tuple[str, ...] | None:
-    if value is None:
-        return None
-    if not isinstance(value, Sequence) or isinstance(value, str):
-        raise TypeError(f"{field_name} must be a sequence of strings or null")
-    return _normalize_string_sequence(value, field_name)
-
-
-def _deserialize_keybindings(value: object) -> dict[str, KeybindingValue]:
-    if not isinstance(value, Mapping):
-        raise TypeError("keybindings must be a JSON object")
-    normalized: dict[str, KeybindingValue] = {}
-    for action, keys in value.items():
-        if not isinstance(action, str):
-            raise TypeError("keybinding action ids must be strings")
-        if keys is None:
-            normalized[action] = None
-            continue
-        if isinstance(keys, str):
-            normalized[action] = keys
-            continue
-        normalized[action] = _normalize_string_sequence(keys, f"keybindings.{action}")
-    return normalized
-
-
-def _serialize_keybindings(value: Mapping[str, KeybindingValue]) -> dict[str, object]:
-    serialized: dict[str, object] = {}
-    for action, keys in value.items():
-        serialized[action] = list(keys) if isinstance(keys, tuple) else keys
-    return serialized
-
-
-def _non_negative_small_int(
-    value: object, field_name: str, *, upper_bound: int | None = None
-) -> int:
-    if not isinstance(value, int | float):
-        raise TypeError(f"{field_name} must be a number")
-    coerced = max(0, int(value))
-    if upper_bound is not None:
-        coerced = min(upper_bound, coerced)
-    return coerced
-
-
-def _bounded_int(
-    value: object, field_name: str, *, lower_bound: int, upper_bound: int
-) -> int:
-    if not isinstance(value, int | float):
-        raise TypeError(f"{field_name} must be a number")
-    return max(lower_bound, min(upper_bound, int(value)))
-
-
-def _thinking_budgets(value: object) -> ThinkingBudgetMap | None:
-    if value is None:
-        return None
-    if not isinstance(value, Mapping):
-        raise TypeError("thinking_budgets must be a JSON object or null")
-    normalized: ThinkingBudgetMap = {}
-    for key, item in value.items():
-        if key not in {"minimal", "low", "medium", "high"}:
-            raise ValueError(
-                "thinking_budgets may only contain minimal, low, medium, or high"
-            )
-        if not isinstance(item, int):
-            raise TypeError("thinking_budgets values must be integers")
-        normalized[cast(ThinkingBudgetKey, key)] = item
-    return normalized
-
-
-def _serialize_settings_slice(value: object) -> dict[str, Any]:
-    if isinstance(value, Mapping):
-        return dict(value)
-    return _serialize_dataclass_slice(value)
-
-
-def _serialize_tool_settings(value: object) -> dict[str, Any]:
-    if isinstance(value, ToolSettings):
-        return {
-            "external_tool_policy": _deserialize_external_tool_policy(
-                value.external_tool_policy
-            ),
-            "blocked_tools": list(value.blocked_tools),
-            "ask_tools": list(value.ask_tools),
-            "blocked_substrings": list(value.blocked_substrings),
-            "ask_substrings": list(value.ask_substrings),
-            "blocked_path_substrings": list(value.blocked_path_substrings),
-            "ask_path_substrings": list(value.ask_path_substrings),
-            "approval_mode": _deserialize_headless_approval_mode(value.approval_mode),
-            "approval_reason": value.approval_reason,
-        }
-    if not isinstance(value, Mapping):
-        raise TypeError("tools must be a JSON object")
-    patch = dict(value)
-    if "external_tool_policy" in patch:
-        patch["external_tool_policy"] = _deserialize_external_tool_policy(
-            patch["external_tool_policy"]
-        )
-    for key in (
-        "blocked_tools",
-        "ask_tools",
-        "blocked_substrings",
-        "ask_substrings",
-        "blocked_path_substrings",
-        "ask_path_substrings",
-    ):
-        if key in patch:
-            patch[key] = list(_normalize_string_sequence(patch[key], key))
-    if "approval_mode" in patch:
-        patch["approval_mode"] = _deserialize_headless_approval_mode(
-            patch["approval_mode"]
-        )
-    if "approval_reason" in patch:
-        patch["approval_reason"] = _optional_string(
-            patch["approval_reason"], "approval_reason"
-        )
-    return patch
-
-
-def _serialize_statusline_settings(value: object) -> dict[str, Any]:
-    if isinstance(value, StatusLineControlSettings):
-        return _serialize_dataclass_slice(value)
-    if not isinstance(value, Mapping):
-        raise TypeError("statusline must be a JSON object")
-    normalized: dict[str, Any] = {}
-    for key, item in value.items():
-        field_name = f"statusline.{key}"
-        if key in {
-            "enabled",
-            "model",
-            "workspace",
-            "branch",
-            "session",
-            "permissions",
-            "runtime",
-        }:
-            normalized[key] = _bool_value(item, field_name)
-        elif key in {"queue", "message"}:
-            normalized[key] = _deserialize_statusline_auto_value(item, field_name)
-        elif key == "separator":
-            normalized[key] = _deserialize_statusline_separator(item, field_name)
-        elif key == "style":
-            normalized[key] = _deserialize_statusline_style(item, field_name)
-        else:
-            raise ValueError(f"Unknown statusline setting: {field_name}")
-    return normalized
-
-
-def _serialize_dataclass_slice(value: object) -> dict[str, Any]:
-    if not is_dataclass(value) or isinstance(value, type):
-        raise TypeError("settings slice must be a dataclass or mapping")
-    return dict(asdict(cast(Any, value)))
-
-
-def _drop_removed_settings(
-    patch: Mapping[str, Any],
-    *,
-    scope: SettingsScope,
-    errors: list[SettingsError] | None,
-) -> dict[str, Any]:
-    normalized = dict(patch)
-    for key, message in _REMOVED_SETTING_MESSAGES.items():
-        if key not in normalized:
-            continue
-        normalized.pop(key)
-        if errors is not None:
-            error = ValueError(message)
-            errors.append(SettingsError(scope=scope, message=message, error=error))
-    return normalized
-
-
-def _control_config_to_patch(config: ControlConfig) -> dict[str, Any]:
-    return dict(_CONTROL_CONFIG_CODEC.encode(config))
-
-
-def _apply_dataclass_patch(current: object, patch_value: object, field_name: str):
-    return decode_dataclass_patch(
-        patch_value,
-        current,
-        field_name=field_name,
-    )
-
-
-def _apply_tool_settings_patch(
-    current: ToolSettings, patch_value: object
-) -> ToolSettings:
-    if not isinstance(patch_value, Mapping):
-        raise TypeError("tools must be a JSON object")
-    next_settings = current
-    if "external_tool_policy" in patch_value:
-        next_settings = replace(
-            next_settings,
-            external_tool_policy=_deserialize_external_tool_policy(
-                patch_value["external_tool_policy"]
-            ),
-        )
-    for key in (
-        "blocked_tools",
-        "ask_tools",
-        "blocked_substrings",
-        "ask_substrings",
-        "blocked_path_substrings",
-        "ask_path_substrings",
-    ):
-        if key in patch_value:
-            normalized = _normalize_string_sequence(patch_value[key], key)
-            if key == "blocked_tools":
-                next_settings = replace(next_settings, blocked_tools=normalized)
-            elif key == "ask_tools":
-                next_settings = replace(next_settings, ask_tools=normalized)
-            elif key == "blocked_substrings":
-                next_settings = replace(next_settings, blocked_substrings=normalized)
-            elif key == "ask_substrings":
-                next_settings = replace(next_settings, ask_substrings=normalized)
-            elif key == "blocked_path_substrings":
-                next_settings = replace(
-                    next_settings,
-                    blocked_path_substrings=normalized,
-                )
-            else:
-                next_settings = replace(
-                    next_settings,
-                    ask_path_substrings=normalized,
-                )
-    if "approval_mode" in patch_value:
-        next_settings = replace(
-            next_settings,
-            approval_mode=_deserialize_headless_approval_mode(
-                patch_value["approval_mode"]
-            ),
-        )
-    if "approval_reason" in patch_value:
-        next_settings = replace(
-            next_settings,
-            approval_reason=_optional_string(
-                patch_value["approval_reason"], "approval_reason"
-            ),
-        )
-    return next_settings
-
-
-def _apply_statusline_settings_patch(
-    current: StatusLineControlSettings,
-    patch_value: object,
-) -> StatusLineControlSettings:
-    patch = _serialize_statusline_settings(patch_value)
-    return replace(current, **patch)
-
-
-def _apply_permission_settings_patch(
-    current: PermissionSettings,
-    patch_value: object,
-) -> PermissionSettings:
-    if not isinstance(patch_value, Mapping):
-        raise TypeError("permissions must be a JSON object")
-    unknown = set(patch_value) - {"profile"}
-    if unknown:
-        raise ValueError(f"Unknown permission setting: permissions.{sorted(unknown)[0]}")
-    if "profile" not in patch_value:
-        return current
-    return replace(
-        current,
-        profile=_deserialize_permission_profile(patch_value["profile"]),
-    )
-
-
-def _decode_bool(field_name: str):
-    return lambda raw, current: _bool_value(raw, field_name)
-
-
-def _decode_optional_string(field_name: str):
-    return lambda raw, current: _optional_string(raw, field_name)
-
-
-def _decode_optional_string_tuple(field_name: str):
-    return lambda raw, current: _string_tuple_or_none(raw, field_name)
-
-
-def _decode_string_tuple(field_name: str):
-    def decode(raw: object, current: object) -> tuple[str, ...]:
-        del current
-        if not isinstance(raw, Sequence):
-            raise TypeError(f"{field_name} must be a sequence of strings")
-        return _normalize_string_sequence(raw, field_name)
-
-    return decode
-
-
-def _decode_dataclass(field_name: str):
-    return lambda raw, current: _apply_dataclass_patch(current, raw, field_name)
-
-
-def _encode_optional_tuple(current: object, default: object) -> object:
-    del default
-    return list(cast(Iterable[object], current)) if current is not None else None
-
-
-def _encode_tuple(current: object, default: object) -> object:
-    del default
-    return list(cast(tuple[object, ...], current))
-
-
-def _decode_keybinding_overlay(raw: object, current: object) -> object:
-    return {
-        **cast(Mapping[str, KeybindingValue], current),
-        **_deserialize_keybindings(raw),
-    }
-
-
-def _decode_session_dir(raw: object, current: object) -> object:
-    del current
-    if raw is not None and not isinstance(raw, str):
-        raise TypeError("session_dir must be a string or null")
-    return raw
-
-
-def _decode_package_sources(raw: object, current: object) -> object:
-    del current
-    return _normalize_package_source_sequence(raw)
-
-
-def _encode_package_sources(current: object, default: object) -> object:
-    del default
-    return [
-        _serialize_package_source(source)
-        for source in cast(tuple[PackageSourceConfig, ...], current)
-    ]
-
-
-_CONTROL_CONFIG_CODEC = SchemaConfigCodec(
-    default_factory=ControlConfig,
-    fields=(
-        ConfigFieldSpec(
-            "default_model",
-            decode=lambda raw, current: _deserialize_model_selection(raw),
-            encode=lambda current, default: _serialize_model_selection(
-                cast(ModelSelection | None, current)
-            ),
-        ),
-        ConfigFieldSpec("thinking_level"),
-        ConfigFieldSpec(
-            "steering_mode",
-            decode=lambda raw, current: _deserialize_queue_mode(raw, "steering_mode"),
-        ),
-        ConfigFieldSpec(
-            "follow_up_mode",
-            decode=lambda raw, current: _deserialize_queue_mode(raw, "follow_up_mode"),
-        ),
-        ConfigFieldSpec("theme", decode=_decode_optional_string("theme")),
-        ConfigFieldSpec("system_prompt"),
-        ConfigFieldSpec(
-            "hide_thinking_block",
-            decode=_decode_bool("hide_thinking_block"),
-        ),
-        ConfigFieldSpec("shell_path", decode=_decode_optional_string("shell_path")),
-        ConfigFieldSpec("quiet_startup", decode=_decode_bool("quiet_startup")),
-        ConfigFieldSpec(
-            "shell_command_prefix",
-            decode=_decode_optional_string("shell_command_prefix"),
-        ),
-        ConfigFieldSpec(
-            "npm_command",
-            decode=_decode_optional_string_tuple("npm_command"),
-            encode=_encode_optional_tuple,
-        ),
-        ConfigFieldSpec(
-            "collapse_changelog",
-            decode=_decode_bool("collapse_changelog"),
-        ),
-        ConfigFieldSpec(
-            "enable_install_telemetry",
-            decode=_decode_bool("enable_install_telemetry"),
-        ),
-        ConfigFieldSpec(
-            "enable_skill_commands",
-            decode=_decode_bool("enable_skill_commands"),
-        ),
-        ConfigFieldSpec(
-            "enabled_models",
-            decode=_decode_optional_string_tuple("enabled_models"),
-            encode=_encode_optional_tuple,
-        ),
-        ConfigFieldSpec(
-            "double_escape_action",
-            decode=lambda raw, current: _deserialize_double_escape_action(raw),
-        ),
-        ConfigFieldSpec(
-            "tree_filter_mode",
-            decode=lambda raw, current: _deserialize_tree_filter_mode(raw),
-        ),
-        ConfigFieldSpec(
-            "show_hardware_cursor",
-            decode=_decode_bool("show_hardware_cursor"),
-        ),
-        ConfigFieldSpec(
-            "editor_padding_x",
-            decode=lambda raw, current: _non_negative_small_int(
-                raw,
-                "editor_padding_x",
-                upper_bound=3,
-            ),
-        ),
-        ConfigFieldSpec(
-            "autocomplete_max_visible",
-            decode=lambda raw, current: _bounded_int(
-                raw,
-                "autocomplete_max_visible",
-                lower_bound=3,
-                upper_bound=20,
-            ),
-        ),
-        ConfigFieldSpec(
-            "keybindings",
-            decode=_decode_keybinding_overlay,
-            encode=lambda current, default: _serialize_keybindings(
-                cast(Mapping[str, KeybindingValue], current)
-            ),
-        ),
-        ConfigFieldSpec(
-            "thinking_budgets",
-            decode=lambda raw, current: _thinking_budgets(raw),
-        ),
-        *(
-            ConfigFieldSpec(
-                field_name,
-                decode=_decode_dataclass(field_name),
-                encode=encode_dataclass_diff,
-            )
-            for field_name in (
-                "compaction",
-                "branch_summary",
-                "retry",
-                "images",
-                "terminal",
-                "markdown",
-                "warnings",
-                "method",
-            )
-        ),
-        ConfigFieldSpec(
-            "permissions",
-            decode=lambda raw, current: _apply_permission_settings_patch(
-                cast(PermissionSettings, current), raw
-            ),
-            encode=encode_dataclass_diff,
-            recover_errors=(TypeError, ValueError),
-        ),
-        ConfigFieldSpec(
-            "tools",
-            decode=lambda raw, current: _apply_tool_settings_patch(
-                cast(ToolSettings, current), raw
-            ),
-            encode=encode_dataclass_diff,
-        ),
-        ConfigFieldSpec(
-            "sandbox",
-            decode=_decode_dataclass("sandbox"),
-            encode=encode_dataclass_diff,
-            recover_errors=(TypeError, ValueError),
-        ),
-        ConfigFieldSpec(
-            "statusline",
-            decode=lambda raw, current: _apply_statusline_settings_patch(
-                cast(StatusLineControlSettings, current), raw
-            ),
-            encode=encode_dataclass_diff,
-            recover_errors=(TypeError, ValueError),
-        ),
-        ConfigFieldSpec("session_dir", decode=_decode_session_dir),
-        ConfigFieldSpec(
-            "resource_roots",
-            decode=_decode_string_tuple("resource_roots"),
-            encode=_encode_tuple,
-        ),
-        ConfigFieldSpec(
-            "package_roots",
-            decode=_decode_string_tuple("package_roots"),
-            encode=_encode_tuple,
-        ),
-        ConfigFieldSpec(
-            "package_sources",
-            input_keys=("packages", "package_sources"),
-            output_key="package_sources",
-            decode=_decode_package_sources,
-            encode=_encode_package_sources,
-        ),
-        ConfigFieldSpec(
-            "plugin_sources",
-            decode=_decode_string_tuple("plugin_sources"),
-            encode=_encode_tuple,
-        ),
-        ConfigFieldSpec(
-            "disabled_skills",
-            decode=_decode_string_tuple("disabled_skills"),
-            encode=_encode_tuple,
-        ),
-        ConfigFieldSpec(
-            "disabled_plugins",
-            decode=_decode_string_tuple("disabled_plugins"),
-            encode=_encode_tuple,
-        ),
-    ),
-    removed_fields=_REMOVED_SETTING_MESSAGES,
-    unknown_fields="ignore",
-)
-
-
 class SettingsManager:
     """Manage the standard settings shared by Agent-backed products."""
+
     def __init__(
         self,
         initial: ControlConfig | None = None,
@@ -776,7 +96,7 @@ class SettingsManager:
         self._config = SettingsRuntime(
             ScopedConfigRuntime(
                 LayeredConfig(
-                    codec=_CONTROL_CONFIG_CODEC,
+                    codec=CONTROL_CONFIG_CODEC,
                     layers=(
                         ConfigLayer("global", global_path, persistent=True),
                         ConfigLayer("project", project_path, persistent=True),
@@ -799,14 +119,18 @@ class SettingsManager:
 
     def apply_overrides(self, overrides: Mapping[str, Any] | ControlConfig) -> None:
         patch = (
-            _control_config_to_patch(overrides)
+            control_config_to_patch(overrides)
             if isinstance(overrides, ControlConfig)
             else dict(overrides)
         )
-        patch = _drop_removed_settings(
-            patch,
-            scope="session",
-            errors=self._adapter_errors,
+        patch, removed_messages = prepare_override_patch(patch)
+        self._adapter_errors.extend(
+            SettingsError(
+                scope="session",
+                message=message,
+                error=ValueError(message),
+            )
+            for message in removed_messages
         )
         self._config.update("session", patch)
 
@@ -835,186 +159,94 @@ class SettingsManager:
         self,
         *,
         scope: SettingsScope = "session",
-        default_model: ModelSelection | None | _Unset = _UNSET,
-        thinking_level: ThinkingLevel | _Unset = _UNSET,
-        steering_mode: QueueMode | _Unset = _UNSET,
-        follow_up_mode: QueueMode | _Unset = _UNSET,
-        theme: str | None | _Unset = _UNSET,
-        system_prompt: str | _Unset = _UNSET,
-        hide_thinking_block: bool | _Unset = _UNSET,
-        shell_path: str | None | _Unset = _UNSET,
-        quiet_startup: bool | _Unset = _UNSET,
-        shell_command_prefix: str | None | _Unset = _UNSET,
-        npm_command: Sequence[str] | None | _Unset = _UNSET,
-        collapse_changelog: bool | _Unset = _UNSET,
-        enable_install_telemetry: bool | _Unset = _UNSET,
-        enable_skill_commands: bool | _Unset = _UNSET,
-        enabled_models: Sequence[str] | None | _Unset = _UNSET,
-        double_escape_action: DoubleEscapeAction | _Unset = _UNSET,
-        tree_filter_mode: TreeFilterMode | _Unset = _UNSET,
-        show_hardware_cursor: bool | _Unset = _UNSET,
-        editor_padding_x: float | int | _Unset = _UNSET,
-        autocomplete_max_visible: float | int | _Unset = _UNSET,
-        keybindings: Mapping[str, object] | _Unset = _UNSET,
-        thinking_budgets: ThinkingBudgetMap | None | _Unset = _UNSET,
-        compaction: CompactionSettings | _Unset = _UNSET,
-        branch_summary: BranchSummarySettings | _Unset = _UNSET,
-        retry: RetrySettings | _Unset = _UNSET,
-        images: ImageSettings | _Unset = _UNSET,
-        terminal: TerminalSettings | _Unset = _UNSET,
-        markdown: MarkdownSettings | _Unset = _UNSET,
-        warnings: WarningSettings | _Unset = _UNSET,
-        method: MethodSettings | Mapping[str, object] | _Unset = _UNSET,
-        permissions: PermissionSettings | Mapping[str, object] | _Unset = _UNSET,
-        tools: ToolSettings | Mapping[str, object] | _Unset = _UNSET,
-        sandbox: SandboxSettings | Mapping[str, object] | _Unset = _UNSET,
-        statusline: StatusLineControlSettings | Mapping[str, object] | _Unset = _UNSET,
-        session_dir: str | None | _Unset = _UNSET,
-        resource_roots: Iterable[str] | _Unset = _UNSET,
-        package_roots: Iterable[str] | _Unset = _UNSET,
+        default_model: ModelSelection | None | Unset = UNSET,
+        thinking_level: ThinkingLevel | Unset = UNSET,
+        steering_mode: QueueMode | Unset = UNSET,
+        follow_up_mode: QueueMode | Unset = UNSET,
+        theme: str | None | Unset = UNSET,
+        system_prompt: str | Unset = UNSET,
+        hide_thinking_block: bool | Unset = UNSET,
+        shell_path: str | None | Unset = UNSET,
+        quiet_startup: bool | Unset = UNSET,
+        shell_command_prefix: str | None | Unset = UNSET,
+        npm_command: Sequence[str] | None | Unset = UNSET,
+        collapse_changelog: bool | Unset = UNSET,
+        enable_install_telemetry: bool | Unset = UNSET,
+        enable_skill_commands: bool | Unset = UNSET,
+        enabled_models: Sequence[str] | None | Unset = UNSET,
+        double_escape_action: DoubleEscapeAction | Unset = UNSET,
+        tree_filter_mode: TreeFilterMode | Unset = UNSET,
+        show_hardware_cursor: bool | Unset = UNSET,
+        editor_padding_x: float | int | Unset = UNSET,
+        autocomplete_max_visible: float | int | Unset = UNSET,
+        keybindings: Mapping[str, object] | Unset = UNSET,
+        thinking_budgets: ThinkingBudgetMap | None | Unset = UNSET,
+        compaction: CompactionSettings | Unset = UNSET,
+        branch_summary: BranchSummarySettings | Unset = UNSET,
+        retry: RetrySettings | Unset = UNSET,
+        images: ImageSettings | Unset = UNSET,
+        terminal: TerminalSettings | Unset = UNSET,
+        markdown: MarkdownSettings | Unset = UNSET,
+        warnings: WarningSettings | Unset = UNSET,
+        method: MethodSettings | Mapping[str, object] | Unset = UNSET,
+        permissions: PermissionSettings | Mapping[str, object] | Unset = UNSET,
+        tools: ToolSettings | Mapping[str, object] | Unset = UNSET,
+        sandbox: SandboxSettings | Mapping[str, object] | Unset = UNSET,
+        statusline: StatusLineControlSettings | Mapping[str, object] | Unset = UNSET,
+        session_dir: str | None | Unset = UNSET,
+        resource_roots: Iterable[str] | Unset = UNSET,
+        package_roots: Iterable[str] | Unset = UNSET,
         package_sources: Iterable[PackageSourceConfig | str | Mapping[str, object]]
-        | _Unset = _UNSET,
-        plugin_sources: Iterable[str] | _Unset = _UNSET,
-        disabled_skills: Iterable[str] | _Unset = _UNSET,
-        disabled_plugins: Iterable[str] | _Unset = _UNSET,
+        | Unset = UNSET,
+        plugin_sources: Iterable[str] | Unset = UNSET,
+        disabled_skills: Iterable[str] | Unset = UNSET,
+        disabled_plugins: Iterable[str] | Unset = UNSET,
     ) -> None:
-        patch: dict[str, Any] = {}
-        if not isinstance(default_model, _Unset):
-            patch["default_model"] = _serialize_model_selection(default_model)
-        if thinking_level is not _UNSET:
-            patch["thinking_level"] = thinking_level
-        if steering_mode is not _UNSET:
-            patch["steering_mode"] = _deserialize_queue_mode(
-                steering_mode, "steering_mode"
+        patch = build_settings_patch(
+            AgentSettingsUpdate(
+                default_model=default_model,
+                thinking_level=thinking_level,
+                steering_mode=steering_mode,
+                follow_up_mode=follow_up_mode,
+                theme=theme,
+                system_prompt=system_prompt,
+                hide_thinking_block=hide_thinking_block,
+                shell_path=shell_path,
+                quiet_startup=quiet_startup,
+                shell_command_prefix=shell_command_prefix,
+                npm_command=npm_command,
+                collapse_changelog=collapse_changelog,
+                enable_install_telemetry=enable_install_telemetry,
+                enable_skill_commands=enable_skill_commands,
+                enabled_models=enabled_models,
+                double_escape_action=double_escape_action,
+                tree_filter_mode=tree_filter_mode,
+                show_hardware_cursor=show_hardware_cursor,
+                editor_padding_x=editor_padding_x,
+                autocomplete_max_visible=autocomplete_max_visible,
+                keybindings=keybindings,
+                thinking_budgets=thinking_budgets,
+                compaction=compaction,
+                branch_summary=branch_summary,
+                retry=retry,
+                images=images,
+                terminal=terminal,
+                markdown=markdown,
+                warnings=warnings,
+                method=method,
+                permissions=permissions,
+                tools=tools,
+                sandbox=sandbox,
+                statusline=statusline,
+                session_dir=session_dir,
+                resource_roots=resource_roots,
+                package_roots=package_roots,
+                package_sources=package_sources,
+                plugin_sources=plugin_sources,
+                disabled_skills=disabled_skills,
+                disabled_plugins=disabled_plugins,
             )
-        if follow_up_mode is not _UNSET:
-            patch["follow_up_mode"] = _deserialize_queue_mode(
-                follow_up_mode, "follow_up_mode"
-            )
-        if theme is not _UNSET:
-            patch["theme"] = _optional_string(theme, "theme")
-        if system_prompt is not _UNSET:
-            patch["system_prompt"] = system_prompt
-        if hide_thinking_block is not _UNSET:
-            patch["hide_thinking_block"] = _bool_value(
-                hide_thinking_block, "hide_thinking_block"
-            )
-        if shell_path is not _UNSET:
-            patch["shell_path"] = _optional_string(shell_path, "shell_path")
-        if quiet_startup is not _UNSET:
-            patch["quiet_startup"] = _bool_value(quiet_startup, "quiet_startup")
-        if shell_command_prefix is not _UNSET:
-            patch["shell_command_prefix"] = _optional_string(
-                shell_command_prefix, "shell_command_prefix"
-            )
-        if npm_command is not _UNSET:
-            normalized_npm_command = _string_tuple_or_none(npm_command, "npm_command")
-            patch["npm_command"] = (
-                list(normalized_npm_command)
-                if normalized_npm_command is not None
-                else None
-            )
-        if collapse_changelog is not _UNSET:
-            patch["collapse_changelog"] = _bool_value(
-                collapse_changelog, "collapse_changelog"
-            )
-        if enable_install_telemetry is not _UNSET:
-            patch["enable_install_telemetry"] = _bool_value(
-                enable_install_telemetry, "enable_install_telemetry"
-            )
-        if enable_skill_commands is not _UNSET:
-            patch["enable_skill_commands"] = _bool_value(
-                enable_skill_commands, "enable_skill_commands"
-            )
-        if enabled_models is not _UNSET:
-            normalized_enabled_models = _string_tuple_or_none(
-                enabled_models, "enabled_models"
-            )
-            patch["enabled_models"] = (
-                list(normalized_enabled_models)
-                if normalized_enabled_models is not None
-                else None
-            )
-        if double_escape_action is not _UNSET:
-            patch["double_escape_action"] = _deserialize_double_escape_action(
-                double_escape_action
-            )
-        if tree_filter_mode is not _UNSET:
-            patch["tree_filter_mode"] = _deserialize_tree_filter_mode(tree_filter_mode)
-        if show_hardware_cursor is not _UNSET:
-            patch["show_hardware_cursor"] = _bool_value(
-                show_hardware_cursor, "show_hardware_cursor"
-            )
-        if editor_padding_x is not _UNSET:
-            patch["editor_padding_x"] = _non_negative_small_int(
-                editor_padding_x, "editor_padding_x", upper_bound=3
-            )
-        if autocomplete_max_visible is not _UNSET:
-            patch["autocomplete_max_visible"] = _bounded_int(
-                autocomplete_max_visible,
-                "autocomplete_max_visible",
-                lower_bound=3,
-                upper_bound=20,
-            )
-        if keybindings is not _UNSET:
-            patch["keybindings"] = _serialize_keybindings(
-                _deserialize_keybindings(keybindings)
-            )
-        if thinking_budgets is not _UNSET:
-            patch["thinking_budgets"] = _thinking_budgets(thinking_budgets)
-        if compaction is not _UNSET:
-            patch["compaction"] = _serialize_settings_slice(compaction)
-        if branch_summary is not _UNSET:
-            patch["branch_summary"] = _serialize_settings_slice(branch_summary)
-        if retry is not _UNSET:
-            patch["retry"] = _serialize_settings_slice(retry)
-        if images is not _UNSET:
-            patch["images"] = _serialize_settings_slice(images)
-        if terminal is not _UNSET:
-            patch["terminal"] = _serialize_settings_slice(terminal)
-        if markdown is not _UNSET:
-            patch["markdown"] = _serialize_settings_slice(markdown)
-        if warnings is not _UNSET:
-            patch["warnings"] = _serialize_settings_slice(warnings)
-        if method is not _UNSET:
-            patch["method"] = _serialize_settings_slice(method)
-        if permissions is not _UNSET:
-            patch["permissions"] = _serialize_settings_slice(permissions)
-        if tools is not _UNSET:
-            patch["tools"] = _serialize_tool_settings(tools)
-        if sandbox is not _UNSET:
-            patch["sandbox"] = _serialize_settings_slice(sandbox)
-        if statusline is not _UNSET:
-            patch["statusline"] = _serialize_statusline_settings(statusline)
-        if session_dir is not _UNSET:
-            patch["session_dir"] = session_dir
-        if not isinstance(resource_roots, _Unset):
-            patch["resource_roots"] = list(
-                _normalize_string_sequence(resource_roots, "resource_roots")
-            )
-        if not isinstance(package_roots, _Unset):
-            patch["package_roots"] = list(
-                _normalize_string_sequence(package_roots, "package_roots")
-            )
-        if not isinstance(package_sources, _Unset):
-            patch["packages"] = [
-                _serialize_package_source(source)
-                for source in _normalize_package_source_sequence(
-                    list(package_sources), "package_sources"
-                )
-            ]
-        if not isinstance(plugin_sources, _Unset):
-            patch["plugin_sources"] = list(
-                _normalize_string_sequence(plugin_sources, "plugin_sources")
-            )
-        if not isinstance(disabled_skills, _Unset):
-            patch["disabled_skills"] = list(
-                _normalize_string_sequence(disabled_skills, "disabled_skills")
-            )
-        if not isinstance(disabled_plugins, _Unset):
-            patch["disabled_plugins"] = list(
-                _normalize_string_sequence(disabled_plugins, "disabled_plugins")
-            )
+        )
 
         layer = scope if scope in {"global", "project"} else "session"
         self._config.update(layer, patch)
@@ -1295,7 +527,7 @@ class SettingsManager:
         *,
         scope: PermissionProfileScope = "session",
     ) -> None:
-        resolved_id = _deserialize_permission_profile(profile_id)
+        resolved_id = permission_profile(profile_id).profile_id
         if not self._permission_profile_ceiling.allows(resolved_id):
             raise PermissionError(
                 self._permission_profile_ceiling.reason
@@ -1348,7 +580,7 @@ class SettingsManager:
             scope=scope,
             tools=replace(
                 self._settings.tools,
-                external_tool_policy=_deserialize_external_tool_policy(policy),
+                external_tool_policy=policy,
             ),
         )
 
@@ -1398,7 +630,7 @@ class SettingsManager:
         *,
         scope: SettingsScope = "project",
     ) -> bool:
-        candidate = _deserialize_package_source(source)
+        candidate = decode_package_source(source)
         current_sources = self._settings.package_sources
         next_sources = _with_package_source(current_sources, candidate)
         self.update_settings(scope=scope, package_sources=next_sources)
